[#82482] zpl: profilers: memory: memory_profiler: Fix thread ID casting

+++ zpl/profilers/memory/memory_profiler.c
@@ -27,7 +27,7 @@ void zpl_profile_stack(const struct k_thread *thread, void *user_data)
 
 	zpl_emit_memory_for_thread_event(ZPL_STACK, thread->stack_info.start,
 			thread->stack_info.size - unused, unused,
-			(uint32_t)thread);
+			(uint32_t)(uintptr_t)thread);
 }
 
 void zpl_profile_heap(void)
